requirements.txt
asgiref==3.7.2
boto3==1.29.3
botocore==1.32.3
certifi==2023.7.22
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.6
django-cors-headers==4.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
PyMySQL==1.1.0
openapi-codec==1.3.2
packaging==23.2
phonenumbers==8.13.23
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
